Fixed recipes not showing on config change in mainactivity

+++ app/src/main/java/xyz/bnayagrawal/android/bakingapp/model/Ingredient.java
@@ -14,7 +14,7 @@ public final class Ingredient implements Parcelable {
     private String measure;
     private String ingredient;
 
-    public Ingredient(int quantity, String measure, String ingredient) {
+    public Ingredient(double quantity, String measure, String ingredient) {
         this.quantity = quantity;
         this.measure = measure;
         this.ingredient = ingredient;
